requirements.txt
garminconnect>=0.2.19
anthropic>=0.40.0
python-dotenv>=1.0.0
keyring>=25.0.0
rich>=13.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
pystray>=0.19.0
Pillow>=10.0.0
